Quality of Life Changes
1. Complete Navbar
2. Add Support for Buttons in WhyHackUMass
