+++ .eslintrc.json
@@ -1,5 +1,6 @@
 {
   "extends": [
+    "next/babel",
     "next/core-web-vitals"
   ]
 }
