+++ app/layout.js
@@ -6,6 +6,19 @@ const inter = Inter({ subsets: ["latin"] });
 export const metadata = {
   title: "HackUMass XII - An outpost for your craziest ideas",
   description: "HackUMass is an outpost for your craziest ideas. November 8 - 10, 2024 at UMass Amherst.",
+  metadataBase: new URL('http://localhost:3000'),
+  // openGraph: {
+  //   title: 'Title webtsite',
+  //   description: 'this is the desciption',
+  //   image: 'url/image.png'
+  // },
+  // twitter: {
+  //   card: 'summary_large_image',
+  //   site: '@eMartiiin94',
+  //   title: 'Title webtsite',
+  //   description: 'this is the desciption',
+  //   image: 'url/image.png'
+  // }
 };
 
 export default function RootLayout({ children }) {
