+++ components/Navbar/navbar.css
@@ -1,18 +1,3 @@
-.navbar {
-    background: #FFF6EC;
-    height: 150px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.2vw;
-    position: sticky;
-    top: 0;
-    z-index: 999;
-    box-shadow: 0 4px 10px 3px rgba(0, 0, 0, 0.3);
-}
-
-
 .navbar-landing {
     background: transparent;
     height: 148px;
@@ -55,17 +40,20 @@
 
 .nav-menu {
     display: grid;
-    grid-template-columns: repeat(5, auto);
-    grid-gap: 50px;
+    grid-template-columns: repeat(9, auto);
+    grid-gap: 20px;
     list-style: none;
     text-align: center;
     width: max-content;
-    justify-content: end;
+    justify-content: center;
+    align-items: center;
     margin-right: 2rem;
 }
 
 .nav-item {
-    font-size: 0.938vw;
+    font-family: 'Inter', sans-serif;
+    font-weight: 500;
+    font-size: 0.94vw;
 }
 
 .nav-links {
